Eric Sunshine fixed compilation error due to ambiguity in invocation of iCelDataBuffer::Add() invoked from celPcActorMove::GetPersistentData().

+++ cel/trunk/plugins/propclass/actormove/actormove.cpp
@@ -485,7 +485,7 @@ csPtr<iCelDataBuffer> celPcActorMove::GetPersistentData (
   }
   else if (spr3d)
   {
-    databuf->Add (spr3d->GetCurFrame ());
+    databuf->Add ((int32)spr3d->GetCurFrame ());
     databuf->Add (spr3d->GetCurAction ()->GetName ());
   }
